Separate the basic front project structure and the basic  react one
